Reformatted the database to fit with the way it was discussed at the meeting.
Refactored all the functions and files that dealt with the database to fit the new Schemas.

Created 2 new routes for application and staff calls.
Added handlers for these routes to communicate with.

Removed command line program code - this will be run from a separate project.

Changed the phoneFormatter and mappers to simply return the object instead of passing it in a callback.

Added a missing closing tag in twiml generator.

Redid some of the tests now that many functions were modified.