Added the add handler for staff.
Will check if the app passed in exists.
Will then verify inputs and format them correctly.
Will then check if a staff member with that name exists
If it doesnt, it will create it and add it to the application.
If it does, it will grab it and add it to the application