Added more to the segments API.  Also began testing for it.  Get and Post are both tested.  Put and Delete are added but not tested.
Also updated segment validator and it's test.  It was assuming the segment's staffers were a staff object but they're actually simply an ObjectID.

Fixed the issue with dates in angular. I was forced to convert the moment date to utc every time we made one to avoid javascripts inherent conversion of dates to local time, which caused the days displayed to be off by 1.