Added passport and a helper module to work with google's oauth
Using their example, this was easily implemented in using sessions.

Made some changes to the routes to verify authentication for the UI.  I tried doing it with the REST API as well, and it worked apart from the tests then failing. For now the REST API is not protected until we can find a way to mock the login.
Added new routes as well.
Edited the layout page to display the users login name and a dropdown allowing them to log out.
New page was added saying there was an authentication error if you arent a bandwidth member.